requirements.txt
streamlit
langchain
langchain-experimental
llama-index
weaviate-client
openai
sqlalchemy
debugpy
openpyxl
PyPDF2
pypdf
docx2txt
PyCryptodome
graphviz==0.20.1
networkx==3.1
beautifulsoup4==4.12.2
